refactor(agent_registry/registry): merge inline tests into tests.rs (#1266)
Inline test module in registry.rs is moved into the existing
agent_registry/tests.rs as a nested mod registry_inline {} submodule
to scope local helpers (temp_registry_dir, make_entry) that would
otherwise collide with existing helpers (test_registry, sample_entry).

- src/agent_registry/registry.rs: 446 → 225 LOC (under 400-line rule)
- src/agent_registry/tests.rs: 141 → 367 LOC
- Wraps extracted block in nested mod registry_inline {}
- Adds use std::path::{Path, PathBuf} inside nested mod (not inherited)

Pure refactor: zero logic changes, all 25 tests pass.

Closes 6/70 of #1266 (this loop iteration), 14 PRs cumulative since #1285.
